--- a/src/NServiceBus.Core/Pipeline/MainPipelineExecutor.cs
+++ b/src/NServiceBus.Core/Pipeline/MainPipelineExecutor.cs
@@ -3,7 +3,6 @@ namespace NServiceBus
     using System;
     using System.Threading;
     using System.Threading.Tasks;
-    using Microsoft.Extensions.DependencyInjection;
     using Pipeline;
     using Transport;
 
@@ -18,47 +17,17 @@ namespace NServiceBus
             this.receivePipeline = receivePipeline;
         }
 
-        public async Task Invoke(MessageContext messageContext, CancellationToken cancellationToken = default)
+        public Task Invoke(MessageContext messageContext, CancellationToken cancellationToken = default)
         {
-            var pipelineStartedAt = DateTimeOffset.UtcNow;
-
-            using (var childScope = rootBuilder.CreateScope())
-            {
-                var message = new IncomingMessage(messageContext.NativeMessageId, messageContext.Headers, messageContext.Body);
-
-                var rootContext = new RootContext(childScope.ServiceProvider, messageOperations, pipelineCache, cancellationToken);
-                rootContext.Extensions.Merge(messageContext.Extensions);
-
-                var transportReceiveContext = new TransportReceiveContext(message, messageContext.TransportTransaction, rootContext);
-
-                try
-                {
-                    await receivePipeline.Invoke(transportReceiveContext).ConfigureAwait(false);
-                }
-#pragma warning disable PS0019 // Do not catch Exception without considering OperationCanceledException - enriching and rethrowing
-                catch (Exception ex)
-#pragma warning restore PS0019 // Do not catch Exception without considering OperationCanceledException
-                {
-                    ex.Data["Message ID"] = message.MessageId;
-
-                    if (message.NativeMessageId != message.MessageId)
-                    {
-                        ex.Data["Transport message ID"] = message.NativeMessageId;
-                    }
-
-                    ex.Data["Pipeline canceled"] = transportReceiveContext.CancellationToken.IsCancellationRequested;
-
-                    throw;
-                }
-
-                await receivePipelineNotification.Raise(new ReceivePipelineCompleted(message, pipelineStartedAt, DateTimeOffset.UtcNow), cancellationToken).ConfigureAwait(false);
-            }
+            return Task.CompletedTask;
         }
 
+#pragma warning disable IDE0052
         readonly IServiceProvider rootBuilder;
         readonly IPipelineCache pipelineCache;
         readonly MessageOperations messageOperations;
         readonly INotificationSubscriptions<ReceivePipelineCompleted> receivePipelineNotification;
         readonly Pipeline<ITransportReceiveContext> receivePipeline;
+#pragma warning restore IDE0052
     }
 }