I've been reflecting for years how I can make the receive pipeline faster. Significantly faster! Today I finally found the solution. 

| Method     | Mean          | Error       | StdDev      | Ratio         | RatioSD  | Gen 0            | Gen 1 | Gen 2 | Allocated |
|------------|---------------|-------------|-------------|---------------|----------|------------------|-------|-------|-----------|
| **Before** | **7000.15 s** | **0.907 h** | **0.757 5** | **1.00**      | **0.00** | **1005389.0110** | **-** | **-** | **1 TB**  |
| After      | 17.74 ns      | 0.226 ns    | 0.278 ns    | 0.00000304587 | 0.00     | -                | -     | -     | -         |


[Benchmark used](https://youtu.be/xCorJG9mubk?t=19)

- :scissors: Pipeline Execution
- :scissors: Childbuilder
- :scissors: Outbox Storage
- :scissors: Synchronized Storage
- :scissors: Saga Storage
- :scissors: Handler Execution