@danielmarbach 2022-04-01
After @jbogard threw in [that async challenge](https://github.com/Particular/NServiceBus/pull/6332#discussion_r840632692) I'm coming to terms the technical complexity of this async machinery here is too high. Closing. Maybe I can come up with something better in a year's time-frame